+++ Guardfile
@@ -1,0 +1,5 @@
+guard :shell do
+  watch /\.lua/ do |m|
+    `busted impl_test.lua`
+  end
+end
